==> notebooks/01_load_and_validate_raw_data.ipynb <==
import pathlib
import requests

def download_one_file_of_raw_data(year, month):  # given a year and month downloads that data
    # this is the url that immediately downloads the data on the nyc-tax-site, pasing placeholders year and month
    URL = f'https://d37ci6vzurychx.cloudfront.net/trip-data/yellow_tripdata_{year}-{month:02d}.parquet'
    response = requests.get(URL) # sends http-get-request to that url, returns a response object which contains the servers response to request which is the taxi-data

    if response.status_code == 200: # if response was successful
        path = f'../data/raw/rides_{year}-{month:02d}.parquet'  # contructs a file path inside data folder, raw-folder, specifying file name yaer-month.extension
        open(path, "wb").write(response.content) # open the file specified by the path and write the contents of the response object which is data into it
        return path
    else:
        raise Exception(f"{URL} is not available")


download_one_file_of_raw_data(year=2022, month=1)  # year-2022 january

import pandas as pd
# year-2022 january
file_path = "../data/raw/rides_2022-01.parquet"  # path of data-file in our project, that we pulled and stored
rides = pd.read_parquet(file_path)  # since data-files are stored as parquet files, reads parquet file given path and returns pandas-dataframe
rides.head(20) # print first 20-rows of data

rides = rides[["tpep_pickup_datetime", "PULocationID"]]  # selects these two columns from dataframe-rides and sets rides equal to only those columns

rides.rename(columns={
    "tpep_pickup_datetime": "pickup_datetime",
    "PULocationID": "pickup_location_id",
}, inplace=True)
# renames the columns-title

rides.head(20)


# VALIDATION STEP
rides["pickup_datetime"].describe() # gives statistics on the values of that column
# the MIN in pickup-datetime columns is 2008, we only want 2022-jan data
# the MAX is the 5th month which is above the 2022-jan data range
# clearly there is data that shoudn't be there

# exclude data that is prior to beginning of jan. jan-feb
rides = rides[rides.pickup_datetime >= "2022-01-01"] # keep data after 1st of 2022-jan 
rides = rides[rides.pickup_datetime < "2022-02-01"] # keep data beofre 1st of 2022-feb

rides.describe()  # rerun describe, all rows in valid dates


# saves validated-dataframe to a parquet file in specified path
rides.to_parquet("../data/transformed/validated_rides_2022_01.parquet")
