==> notebooks/02_transform_raw_data_into_ts_data.ipynb <==
import pandas as pd

rides = pd.read_parquet("../data/transformed/validated_rides_2022_01.parquet") # reads in parquet-file in specified path and returns its data in a dataframe

rides.head(10)  
# each row in dataframe represents a ride

# add column that is the rounded hour, because we want to work with time series data at an hourly frequency
# rides[new-col] = rides get the datetime columns and round the hour part of it. .dt acess the datetime properties of the pickup_datetime columns
rides["pickup_hour"] = rides["pickup_datetime"].dt.floor("H")
rides

# group/count the number of rides per location_ID per pickup_hour
# groupby(): given the rides-dataframe, groups the 2 columns which means all the rows with the same pickup_hour and pickup_location_id will be grouped together
# size(): counts the number rows/rides ine ach group, this produces a series wher ehte index is the grouped columns, and the values are the counts of rides in each group, 
# reset-index(): converts the series back into a dataframe and resets the index
app_rides = rides.groupby(["pickup_hour","pickup_location_id"]).size().reset_index()
app_rides.rename(columns={0: "num_of_rides",}, inplace=True)  # rename the new column to the number of rides per location and hour
app_rides

# for instances of locaiton and hour where there were no rides, they do not appear above, so we want to place them with zeros
from tqdm import tqdm  # adds loading-bar for for-loops

def add_missing_slots(agg_rides):
    location_ids = agg_rides['pickup_location_id'].unique() # returns array of unqiue pickup-location-ids
    # date-time-index object representing complete range of hourly time slots from min to max pickup-hour in agg-rides
    full_range = pd.date_range(agg_rides["pickup_hour"].min(), agg_rides["pickup_hour"].max(), freq="H")
    # an empty df to store the final result. 
    output = pd.DataFrame()
    
    # iterate throughn every unqiue location-id
    for location_id in tqdm(location_ids):
        # if df.attribute is equal to cur-location-id, loc selects teh specified columns, aslong as the conditions is met
        # first filters rows by if its equal to cur-location-id, then selects the two columns 
        agg_rides_i = agg_rides.loc[agg_rides.pickup_location_id == location_id, ["pickup_hour","num_of_rides"]]  # return is the filtered data-frame

        agg_rides_i.set_index("pickup_hour", inplace=True) # sets the pickup-hour-col as the index of the df-agg-rides-i
        agg_rides_i.index = pd.DatetimeIndex(agg_rides_i.index)  #setting the index-aattribute equal to DTI-obj
        agg_rides_i = agg_rides_i.reindex(full_range, fill_value=0) # fills any missing values in agg-rides-i with 0
        agg_rides_i["pickup_location_id"] = location_id  # adding new column equal to cur-location-id

        output = pd.concat([output, agg_rides_i]) # mergs cur-agg-rides-i

    output = output.reset_index().rename(columns={"index":"pickup_hour"}) # modifies output-df by restting its index and renaming the newly created column

    return output
     



agg_rides_all_slots = add_missing_slots(app_rides)   

# plot time series data to detect any bugs that may be in your data-processing. Before doing machine learning it is recommended that you get really familier with your data.
from typing import Optional, List
import plotly.express as px

def plot_rides(
    rides: pd.DataFrame,
    locations: Optional[List[int]] = None
    ):
    """
    Plot time-series data
    """
    rides_to_plot = rides[rides.pickup_location_id.isin(locations)] if locations else rides

    fig = px.line(
        rides_to_plot,
        x="pickup_hour",
        y="num_of_rides",
        color='pickup_location_id',
        template='none',
    )

    fig.show()
     
 
plot_rides(agg_rides_all_slots, locations=[43])  # what is this 43?

# saving data-frame as a parquet file in specified path, this is transformed data
agg_rides_all_slots.to_parquet("../data/transformed/ts_data_2022_01.parquet")
