==> notebooks/04_transform_raw_data_into_features_into_targets.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys
import os
import importlib
script_dir = os.getcwd()  
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))
import src.data
importlib.reload(src.data) # Reload the module to reflect changes

from src.data import load_raw_data

# passing no months will download all months, this downloads all raw-data for specified year and months
rides = load_raw_data(year=2022)  
rides


from src.data import transform_raw_data_into_ts_data

# rounds the hours, and groups the pikcup-hour and pickup-location-id column together, adds zero where there are not rides for a particular location and hour
ts_data = transform_raw_data_into_ts_data(rides)
ts_data

from src.data import transform_ts_data_into_features_and_target

# convverts ts-data into features/targets-dataframe, givena  single location if ith-hour is the target, the previous 12-hours are the features, so basically does this through slicing
features, targets = transform_ts_data_into_features_and_target(ts_data, input_seq_len=24*28*1, step_size=24)

print(f"features: {features.shape=}")
print(f"targets: {targets.shape=}")

tabular_data = features   # set tabular-data-var equa to the features-df
tabular_data["target_rides_next_hour"] = targets  # add a clolumn which is equal to the targets-df
 
from src.paths import TRANSFORMED_DATA_DIR
tabular_data.to_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")  # convert dataframe into parquet-file and save it in specified path
