==> notebooks/05_visualize_training_data.ipynb <==
%reload_ext autoreload
%autoreload 2

import sys
import os
import importlib
script_dir = os.getcwd()  
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))
import src.data
importlib.reload(src.data) 

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

features_and_target = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")  # load-in the training-data parquet file
features_and_target

# split the train-data into features/targets
features = features_and_target.drop(columns=["target_rides_next_hour"])   # train-data-df drop the target-column to get only the features
targets = features_and_target["target_rides_next_hour"]                   # only select teh target column to get only the targets
print(f"features: {features.shape=}")
print(f"targets: {targets.shape=}")
features["rides_previous_672_hour"][0]

# WARNING: the ploting of training-tabular-data is not working
from src.plot import plot_one_sample, hello

plot_one_sample(features, targets, 0)
