==> notebooks/06_baseline_model.ipynb <==
%reload_ext autoreload
%autoreload 2
import sys
import os
import importlib
script_dir = os.getcwd()  
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))


import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR


df = pd.read_parquet(TRANSFORMED_DATA_DIR / "tabular_data.parquet")
df

from datetime import datetime
from src.data_split import train_test_split # import func from py file

X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date=datetime(2022, 6, 1, 0, 0, 0), target_column_name="target_rides_next_hour")
print(f"{datetime(2022, 6, 1, 0, 0, 0)}")

print(f"X_train: {X_train.shape}") # (examples-in-train, num-features)
print(f"y_train: {y_train.shape}") # (examples-in-train,) 1D beceause only one target
print(f"X_test: {X_test.shape}")   # (examples-in-test, num-features)
print(f"y_test: {y_test.shape}")   # (examples-in-test,) 1D beceause only one target
# examples-in-train != examples-in-test
 


import numpy as np

# class for baseline-model to compare with more complex models in fututre
class BaselineModelPreviousHour:
    """
    Predicts actual deman observed in the last hour
    """
    # takes in training-features-targets, usually model adjusts parameters in this step, but there is nothing to adjust no parameters so it essentially does nothing
    def fit(self, X_train, y_train):
        pass
    
    # takes in test-feautres and returns the demand from preivous hour
    def predict(self, X_test):
        return X_test[f"rides_previous_{7*24}_hour"]  # 7 times 24 so last 7 days

model = BaselineModelPreviousHour()   # create baseline-model-obj
predictions = model.predict(X_test)   # generate predictions base on testing-features
predictions    # print vector of predictions

# MAE for error function, this is regression problem
from sklearn.metrics import mean_absolute_error

test_mae = mean_absolute_error(y_test, predictions)  # built-in-func takes in testing-labels and predictions-vector, outputs mean-absolute-error
print(f"MAE: {test_mae=:.4f}")

# BASELINE MODEL 3
class BaseLineModelLast4Weeks:
    """
    Prediction is actual demand observed at t-7 days, t-21 days, t-28 days
    """

    def fit(self, X_train, y_train):
        pass

    def predict(self, X_test):
        return 0.25 * (
            X_test[f"rides_previous_{7*24}_hour"] + \
            X_test[f"rides_previous_{2*7*24}_hour"] + \
            X_test[f"rides_previous_{3*7*24}_hour"] + \
            X_test[f"rides_previous_{4*7*24}_hour"]
        )


model = BaseLineModelLast4Weeks()
predictions = model.predict(X_test)

test_mae = mean_absolute_error(y_test, predictions)
print(f"MAE: {test_mae=:.4f}")
