==> notebooks/08_lightgbm_model_with_feature_engineering.ipynb <==
%reload_ext autoreload
%autoreload 2
import sys
import os
import importlib
script_dir = os.getcwd()  
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))


import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR

df = pd.read_parquet(TRANSFORMED_DATA_DIR/ "tabular_data.parquet")
df

from datetime import datetime
from src.data_split import train_test_split # import func from py file

X_train, y_train, X_test, y_test = train_test_split(df, cutoff_date=datetime(2022, 6, 1, 0, 0, 0), target_column_name="target_rides_next_hour")

print(f"X_train: {X_train.shape}") # (examples-in-train, num-features)
print(f"y_train: {y_train.shape}") # (examples-in-train,) 1D beceause only one target
print(f"X_test: {X_test.shape}")   # (examples-in-test, num-features)
print(f"y_test: {y_test.shape}")   # (examples-in-test,) 1D beceause only one target
# examples-in-train != examples-in-test
# x_train
# y_train

# FEATURE ENGINEERING
def average_rides_last_4_weeks(X):  # given X-dataframe-train-data
    # create new column avr-rides-4-weeks = mean of the four columns of last 4 weeks
    X["average_rides_last_4_weeks"] = 0.25*(
        X[f"rides_previous_{7*24}_hour"] + \
        X[f"rides_previous_{2*7*24}_hour"] + \
        X[f"rides_previous_{3*7*24}_hour"] + \
        X[f"rides_previous_{4*7*24}_hour"] 
    )

    return X

from sklearn.preprocessing import FunctionTransformer

# wraps the func-aver-rides-last-4-weeks so it can be used in scikit learn pipeline
add_feature_average_rides_last_4_weeks = FunctionTransformer(average_rides_last_4_weeks, validate=False)

# applying the function-transformer to the X-train-data-frame adding in the new column
add_feature_average_rides_last_4_weeks.fit_transform(X_train)

from sklearn.base import BaseEstimator, TransformerMixin

# create custom Transformer-class that inherits from these two scikit learn classes
class TemporalFeaturesEngineer(BaseEstimator, TransformerMixin):

    def fit(self, X, y=None):
        return self
    
    # takes in X-df
    def transform(self, X=None):

        X_ = X.copy()  # create copy of X-df
        # perform feature-engineering, by creating new column-hour which is hour of the pickup-hour-column
        X_["hour"] = X_["pickup_hour"].dt.hour
        X_["day_of_week"] = X_["pickup_hour"].dt.dayofweek  # set new column day-of-week for each row equal to that rows pickup-hour-columns day part

        return X_.drop(columns=["pickup_hour"])  # drop this column

add_temporal_features = TemporalFeaturesEngineer() # create custom-transformer-obj
add_temporal_features.fit_transform(X_train)  # applying function transform which which adds in the two new columms

import lightgbm as lgb  # model-library

from sklearn.pipeline import make_pipeline  # import func that creates a pipeline

# creating pipeline passing sequence of transformations on X, and the model
pipeline = make_pipeline(add_feature_average_rides_last_4_weeks, add_temporal_features, lgb.LGBMRegressor())
pipeline.fit(X_train, y_train) # method applies each step of the pipeline and trains the model



from sklearn.metrics import mean_absolute_error

predictions = pipeline.predict(X_test)
test_mae = mean_absolute_error(y_test, predictions)  # built-in-func takes in testing-labels and predictions-vector, outputs mean-absolute-error
print(f"MAE LGBM-BOOST: {test_mae=:.4f}")
