==> notebooks/11_backfill_feature_store.ipynb <==
%reload_ext autoreload
%autoreload 2
import sys
import os
import importlib
script_dir = os.getcwd()  
sys.path.append(os.path.abspath(os.path.join(script_dir, '..')))

HOPSWORKS_PROJECT_NAME = "taxi_demand_pravachan"

import os
from dotenv import load_dotenv
from src.paths import PARENT_DIR

load_dotenv(PARENT_DIR / ".env")  # load the .env environment variable file by passing in its path
HOPSWORKS_API_KEY = os.environ["HOPSWORKS_API_KEY"]  # extract the api-key-var from the file



from datetime import datetime
import pandas as pd
from src.data import load_raw_data  # func

from_year = 2022  # starting year
to_year = datetime.now().year   # ending year which is now
print(f"Downloading raw data from {from_year} to {to_year}")
rides = pd.DataFrame()  # initlize empty dataframe to store all rides in date range
# iterate every year from    starting-year to inclusive ending year
for year in range(from_year, to_year+1):
    # calling load-raw-data-func passing in cur-year, downloads raw-data of all months of cur-year from website and validates it and returns that data-frame
    rides_one_year = load_raw_data(year)
    # add the rides-data-frame for cur year to the all-rides-df
    rides = pd.concat([rides, rides_one_year])

print(f"{len(rides)}")   # number of rows of data, each row represents a ride

from src.data import transform_raw_data_into_ts_data

ts_data = transform_raw_data_into_ts_data(rides)  # transforms rides-df-raw into time-series data-frame is returned

import hopsworks  

# creating hopsworks-project by passing project name and api-key
project = hopsworks.login(project=HOPSWORKS_PROJECT_NAME, api_key_value=HOPSWORKS_API_KEY)

# hopsworks-project.get-feature-store() gest pointer to the feature store
feature_store = project.get_feature_store()

FEATURE_GROUP_NAME = "time_series_hourly_feature_group"
FEATURE_GROUP_VERSION =  3  # name and version of feature group

# creating feature-group-obj feature-store-obj.create()
# primary-key: unique identifier for each row in data
# event-time: is the column that stores the tiemstamp behind those events
feature_group = feature_store.get_or_create_feature_group(
    name=FEATURE_GROUP_NAME,
    version=FEATURE_GROUP_VERSION,
    description="Time-series data at hourly frequency",
    primary_key=["pickup_location_id", "pickup_hour"],
    event_time="pickup_hour",
)

# save data to feature-group
feature_group.insert(ts_data, write_options={"wait_for_job":False})
